Update entry points and report generation for folder-based splits
